internal: move indent function to internal package
Indent function currently is part of the linkerd component, this is also
needed in the velero component to indent the credentials passed by the
user.

This commit then moves the function to the `internal` package and its
corresponding unit test.
